requirements.txt
faiss-cpu
numpy==1.26.4
pillow
tqdm
torch
torchvision
open-clip-torch
opencv-python
transformers
sentencepiece
streamlit
